[Fix #160468381] App restructure Restructures the app and versions the files for better use
