+++ API_V1/app/v1/models/manFoods.py
@@ -1,7 +1,7 @@
 from app import api
-from app.models.manOrders import food_items
-from app.utility.validFood import FoodDataValidator
-from app.utility.validOrder import OrderDataValidator
+from app.v1.models.manOrders import food_items
+from app.v1.utility.validFood import FoodDataValidator
+from app.v1.utility.validOrder import OrderDataValidator
 
 
 """
